% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-minor-tickmarks.r
\name{add_minor_tickmarks}
\alias{add_minor_tickmarks}
\title{Add minor unlabelled tickmarks to an existing plot}
\usage{
add_minor_tickmarks(
  x_tick_start = 0,
  x_tick_by = 2,
  x_tick_end = NULL,
  y_tick_start = 0,
  y_tick_by = 0.01,
  y_tick_end = NULL
)
}
\value{
adds minor tickmarks to an existing plot
}
\description{
Add sensible smaller (unlabelled) tickmarks to both axes of an existing
plot. Adapted from \code{pacea::add_tickmarks()}. Is exported but unlikely to be needed externally.
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
