% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-density.R
\name{calc_density}
\alias{calc_density}
\title{For a vector of values, calculate the equal and HDI intervals and the kernel
density with corresponding equal and HDI intervals}
\usage{
calc_density(dat_mcmc, intervals = NULL)
}
\description{
For a vector of values, calculate the equal and HDI intervals and the kernel
density with corresponding equal and HDI intervals
}
\examples{
\dontrun{
one_year_mcmc <- pull(hake_recruitment_mcmc, 2021)
calc_density(one_year_mcmc)
}
}
\author{
Andrew Edwards
}
