% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-density-many-years.R
\name{calc_density_many_years}
\alias{calc_density_many_years}
\title{For a tibble of MCMC samples as rows and some quantity in each column
(assuming years, could generalise), calculate the equal and HDI intervals and the kernel
densities for each column.}
\usage{
calc_density_many_years(dat_mcmc, ...)
}
\arguments{
\item{...}{arguments to pass to \code{calc_density()}, that then get
passed onto \code{create_intervals_density()} and then \code{density()}}

\item{density}{if TRUE (the default) then use the density approach for the HDI
calculation, rather than the \code{hdi()} default of just the sample values.}
}
\value{
list object of \code{res_all_years} which contains \code{year} plus
\code{intervals} and \code{dens} for that year, plus \code{intervals_all_years} which
contains all the intervals in one tibble with \code{year} as the first column.
}
\description{
TODO \code{n=1e5} is hardwired in here
}
\examples{
\dontrun{
TODOone_year_mcmc <- dplyr::pull(hake_recruitment_mcmc, `2021`)
TODOcalc_density(one_year_mcmc)
}
}
\author{
Andrew Edwards
}
