% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-hake-forecast-mcmc.R
\name{create_hake_forecast_mcmc}
\alias{create_hake_forecast_mcmc}
\title{Extract relative spawning biomass calculations for forecasts for a given
catch alternative.}
\usage{
create_hake_forecast_mcmc(
  catch_alternative = "06-350000",
  forecast_year_dir = "2027",
  model_path = NULL,
  forecast_yrs = 2024:2027
)
}
\arguments{
\item{catch_alternative}{which catch alternative (with index) to use, from
Table g}

\item{forecast_year_dir}{which of the forecast years directories to use}

\item{model_path}{where to find the model Stock Synthesis results, gets set to
Andy's directory if NULL}

\item{forecast_yrs}{the forecast years}
}
\value{
tibble with each row an MCMC sample and characters as columns for
the \code{forecast_yrs} to be returned
}
\description{
Need the full MCMC Stock Synthesis results in the correct folder, as the
assessment build does not automatically make .rds files that include MCMC
results for all model runs (it does for the base model). Not needed for
users, used to create \code{hake_relative_biomass_mcmc_forecast} in \code{data-raw/hake-forecasts.R}.
}
\examples{
\dontrun{
# See data-raw/hake-forecasts.R; not needed by users
}
}
\author{
Andrew Edwards, with some adapted from \code{hake::load_forecasts()}
originally written by Chris Grandin
}
