% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-intervals.R
\name{create_intervals}
\alias{create_intervals}
\title{Make \code{create_intervals()} use the right method depending on class of input}
\usage{
create_intervals(dat, ...)
}
\description{
Based on \code{plot()}. The functions are \code{create_itervals.numeric()} and \code{create_intervals.data.frame()}.
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
