% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-intervals-density.R
\name{create_intervals_density}
\alias{create_intervals_density}
\title{Create credible intervals from a numeric vector using the density approach}
\usage{
create_intervals_density(dat, ...)
}
\arguments{
\item{dat}{a numeric vector consisting of MCMC samples that we want to
summarise with credible intervals}

\item{...}{arguments to pass onto \code{density()}, in particular \code{n} which is
the number of equally-spaced points at which density is to be
estimated.}
}
\value{
tibble of one row (to be consistent with
\code{create_intervals.data.frame()}) with columns \code{median}, \code{2.5} and \code{97.5}
representing the equal-tailed interval, \code{hdi_lower} and \code{hdi_upper}
representing the lower and upper ends of the HDI 95\% interval,
\code{width_equal} and \code{width_hdi} giving the widths of the two types of
intervals, and \code{width_diff} giving the difference between them
(\code{width_equal - width_hdi}, which must be non-negative
because HDI by definition is the narrowest interval).
}
\description{
\if{html}{\out{
 <desc>
}}
}
\examples{
\dontrun{
rec_intervals <- create_intervals(dplyr::pull(hake_recruitment_mcmc, "2021")
}
}
\author{
Andrew Edwards
}
