% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-1.R
\name{figure_1}
\alias{figure_1}
\title{Plot Figure 1 for a specific dimension.}
\usage{
figure_1(
  file = "figure-1.pdf",
  height = 4,
  width = 114/25.4,
  xlim = c(0, 40),
  xlab = "Recruitment (billions of fish)",
  ...
)
}
\description{
Useful for the manuscript, maybe less so for other users, though could be
useful as a template. Generally running in
the vignettes folder then copying over to the write up repository.
}
\examples{
\dontrun{
figure_1()
}
}
\author{
Andrew Edwards
}
