% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-2.R
\name{figure_2}
\alias{figure_2}
\title{Fig 2
\if{html}{\out{<desc>}}}
\usage{
figure_2(file = "figure-2.pdf", height = 4, width = 114/25.4)
}
\description{
Fig 2
\if{html}{\out{<desc>}}
}
\author{
Andrew Edwards
}
