% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-S1.R
\name{figure_s1}
\alias{figure_s1}
\title{Plot Figure S1 as a pdf file with specific dimensions for the manuscript.}
\usage{
figure_s1(
  file = "figure-s1.pdf",
  height = 4,
  width = 114/25.4,
  xlim = c(0, 40),
  xlab = "Recruitment (billions of fish)",
  ...
)
}
\arguments{
\item{file}{filename to save pdf file}

\item{height}{height of pdf file (inches)}

\item{width}{width of pdf file (inches)}

\item{xlim, }{xlab x-axis range and label, as passed onto \code{\link[=plot.numeric]{plot.numeric()}}}

\item{...}{further arguments passed onto \code{\link[=plot.numeric]{plot.numeric()}}}
}
\value{
nothing, but saves the pdf
}
\description{
Useful for the manuscript, maybe less so for other users, though could be
useful as a template. Generally running in
the vignettes folder then copying over to the write up repository. Was
also useful for running the tests.
}
\examples{
\dontrun{
figure_s1()
}
}
\author{
Andrew Edwards
}
