% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hake_recruitment_mcmc}
\alias{hake_recruitment_mcmc}
\title{Help files for data objects that are not committed and so not on GitHub.}
\format{
A tibble with each row representing an MCMC sample, and columns:
\describe{
\item{Virgin:}{presumably gives \code{R0}, TODO double check}
\item{\code{1966}:\code{2027}:}{year for each recruitment}
}
}
\source{
Generated from running \code{data-raw/hake.R}
}
\usage{
hake_recruitment_mcmc
}
\description{
Help files for data objects that are not committed and so not on GitHub.
}
\author{
Andrew Edwards
}
\keyword{datasets}
