% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdiAnalysis.R
\docType{package}
\name{hdiAnalysis}
\alias{hdiAnalysis}
\title{hdiAnalysis package dplyr-related things}
\description{
Taken from pacea, which was adapted from https://github.com/pbs-assess/gfiphc/blob/master/R/gfiphc.R
(see that if get issues with column names within dplyr functions for what to
add). And from
https://github.com/andrew-edwards/sizeSpectra/blob/master/R/sizeSpectra.R for
some tibble related stuff.
}
