% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate-simpsons.R
\name{integrate_simpsons}
\alias{integrate_simpsons}
\title{Integration of a kernel density estimate using Simpson's rule of integration}
\usage{
integrate_simpsons(dens, tol = 1e-08)
}
\arguments{
\item{dens}{kernel density estimate, that has \verb{$x} as the x values with
corresponding \verb{$y} for the density estimates.}

\item{tol}{numeric tolerance for checking that \code{dens$x} are equally spaced;
increase if get an error}
}
\value{
the value of the integral
}
\description{
Based on algorithm in Table 18.4 on p961 of Kreyszig (1993) Advanced
Engineering Mathematics (Seventh Edition). Also the \href{https://en.wikipedia.org/wiki/Simpson's_rule#Simpson.27s_3.2F8_rule}{Composition Simpson's 1/3 rule on Wikipedia},
which is slightly easier notation for implementation, so we follow that here.
}
\details{
Simpson's rule requires an even number of equal subintervals; if there are an
odd number (so even number of points, which is likely the case for density
calculations since the points will be a round number), then we use the
trapezoid rule for the final point, which is likely close to zero anyway
because the density kernel should be reaching far enough away from the bulk
of the density.
}
\examples{
\dontrun{
integrate_simpsons(density(rnorm(8000), n = 1e05))  # n as used in default
  for create_intervals()
}
}
\author{
Andrew Edwards
}
