% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermediate-y.R
\name{intermediate_y}
\alias{intermediate_y}
\title{Calculate an intermediate y value between two (x,y) points.}
\usage{
intermediate_y(x_val, x1, x2, y1, y2)
}
\arguments{
\item{x_val}{numeric value for which to calculate the corresponding y value}

\item{x1, }{x2, y1, y2 numeric value representing co-ordinates (x1, y1) and
(x2, y2) between which we want to calculate the interpolated y value at
\code{x_val}. Require \code{x1 < x2} and \verb{x1 <= x_val <= x2}.}
}
\value{
the intermediate y value
}
\description{
Given x1 and x2 and corresponding y1 and y2, calculate the intermediate
value at \code{x_val} where \verb{x1 <= x_val <= x2}, by a simple linear interpolation.
}
\author{
Andrew Edwards
}
