% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.intervals_density.R
\name{plot.intervals_density}
\alias{plot.intervals_density}
\title{Plot kernel density function of an \code{intervals_density} object, with
tails shaded as specified and optional explanatory lines added}
\usage{
\method{plot}{intervals_density}(
  ints_dens,
  dat = NULL,
  type = "hdi",
  col_main = "blue",
  col_main_text = NULL,
  col_tail = "red",
  main_title = NULL,
  main_title_include = FALSE,
  hdi_horizontal = TRUE,
  col_hdi_horizontal = "darkgrey",
  rug_top = FALSE,
  rug_bottom = FALSE,
  interval_arrows = FALSE,
  y_arrow = 0.098,
  arrowhead_length = 0.2,
  arrowhead_gap = 0,
  col_bars = "black",
  bars_multiplier = 1.5,
  lwd_border = 0.4,
  x_minor_ticks_by = 1,
  y_minor_ticks_by = 0.01,
  ticks_tcl = -0.2,
  show_discontinuity = FALSE,
  discontinuity_multiplier = 2,
  ...
)
}
\arguments{
\item{type}{type of intervals: either \code{hdi} or \code{equal}}

\item{main_title_include}{logical whether to include a main title}

\item{rug_top}{logical whether to show rug at the top for the density values}

\item{rug_bottom}{logical whether to show rug at the bottom for the raw data values}

\item{interval_arrows}{logical whether to show arrows depicting intervals}

\item{y_arrow}{value on y-axis to show the arrows depicting intervals}

\item{col_bars}{colour of the bars showing regions A and B}

\item{bars_multiplier}{numeric multiplier, to nudge the bars higher (value of
1.0 puts them exactly at the minimum density of the ends of the credible
interval)}

\item{show_discontinuity}{logical, if \code{TRUE} then plot the discontinuity in
the HDI that arises (only matters if \code{ints_dens$intervals$warning == TRUE}).}

\item{discontinuity_multiplier}{numeric, y-axis value to plot points showing
discontinuities, multiplies the minimum density of the ends of the
credible interval.}

\item{...}{arguments to pass onto \code{plot()}}

\item{dat_mcmc}{a numeric vector representing an MCMC sample.}

\item{year}{Which year of recruitment (age-0) to plot}

\item{rec_intervals}{result of \code{create_intervals(dat_mcmc)}; is calculated
if not supplied. May be worth supplying so it's not being repeatedly calculated.}
}
\value{
invisible
}
\description{
TODO creates plot on object of class \code{intervals_density}, which is the
result of running \code{create_intervals()} on a vector of values.
}
\examples{
\dontrun{
plot_recruitment_density()
plot_recruitment_density(year = 2021)
}
}
\author{
Andrew Edwards
}
