% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-densitites.R
\name{plot_densities}
\alias{plot_densities}
\title{Calll from plot.intervals_density_list()
\if{html}{\out{<desc>}}}
\usage{
plot_densities(obj, type, quantity = NULL, mfrow = c(2, 2), ...)
}
\description{
Calll from plot.intervals_density_list()
\if{html}{\out{<desc>}}
}
\author{
Andrew Edwards
}
