% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-density.R
\name{plot_density}
\alias{plot_density}
\title{Plot density function (smoothed) of MCMC recruitment for a given year, with tails shaded as specified}
\usage{
plot_density(
  dat_mcmc = one_year_mcmc,
  dens_intervals = NULL,
  year = 2021,
  type = "hdi",
  x_lim = c(0, 40),
  col_main = "blue3",
  col_tail = "red",
  main_title = NULL,
  main_title_include = FALSE,
  x_lab = NULL,
  rug_top = FALSE,
  rug_bottom = FALSE,
  interval_arrows = FALSE,
  y_arrow = 0.095,
  arrowhead_length = 0.2,
  col_bars = "darkgreen",
  bars_multiplier = 1.5,
  ...
)
}
\arguments{
\item{dat_mcmc}{a numeric vector representing an MCMC sample.}

\item{year}{Which year of recruitment (age-0) to plot}

\item{type}{type of intervals: either \code{hdi} or \code{equal}}

\item{main_title_include}{logical whether to include a main title}

\item{rug_top}{logical whether to show rug at the top for the density values}

\item{rug_bottom}{logical whether to show rug at the bottom for the raw data values}

\item{interval_arrows}{logical whether to show arrows depicting intervals}

\item{y_arrow}{value on y-axis to show the arrows depicting intervals}

\item{col_bars}{colour of the bars showing regions A and B}

\item{bars_multiplier}{numeric multiplier, so nudge the bars higher (value of
1.0 puts them exactly at the height of the right-hand end of low credible interval)}

\item{...}{arguments to pass onto \code{plot()}}

\item{rec_intervals}{result of \code{create_intervals(dat_mcmc)}; is calculated
if not supplied. May be worth supplying so it's not being repeatedly calculated.}
}
\value{
invisible
}
\description{
Plot density function (smoothed) of MCMC recruitment for a given year, with tails shaded as specified
}
\examples{
\dontrun{
plot_recruitment_density()
plot_recruitment_density(year = 2021)
}
}
\author{
Andrew Edwards
}
