% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment-density.R
\name{plot_recruitment_density}
\alias{plot_recruitment_density}
\title{DEPRECATED I think for plot-density (so more generic). Plot density function (smoothed) of MCMC recruitment for a given year, with tails shaded as specified}
\usage{
plot_recruitment_density(
  dat_mcmc = hake_recruitment_mcmc,
  rec_intervals = NULL,
  year = 2010,
  type = "hdi",
  x_lim = c(0, 40),
  col_main = "red",
  col_tail = "blue",
  main_title = NULL,
  x_lab = NULL
)
}
\arguments{
\item{dat_mcmc}{a tibble where each row represents an MCMC sample, and each
column is a quantity which we want to summarise with credible intervals.}

\item{rec_intervals}{result of \code{create_intervals(dat_mcmc)}; is calculated
if not supplied. May be worth supplying so it's not being repeatedly calculated.}

\item{year}{Which year of recruitment (age-0) to plot}

\item{type}{type of intervals: either \code{hdi} or \code{equal}}
}
\value{
invisible
}
\description{
DEPRECATED I think for plot-density (so more generic). Plot density function (smoothed) of MCMC recruitment for a given year, with tails shaded as specified
}
\examples{
\dontrun{
plot_recruitment_density()
plot_recruitment_density(year = 2021)
}
}
\author{
Andrew Edwards
}
