% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-table.intervals_density.R
\name{summary_table.intervals_density}
\alias{summary_table.intervals_density}
\title{Print a Markdown summary table of main results from \code{create_intervals()}
output of a single vector}
\usage{
\method{summary_table}{intervals_density}(int_dens, dig = 2)
}
\arguments{
\item{int_dens}{\code{intervals_density} object as output from \code{create_intervals()}}

\item{dig}{number of decimal places to show}
}
\value{
Markdown code (for automatic use straight in an R Markdown document, for instance)
}
\description{
Print a Markdown summary table of main results from \code{create_intervals()}
output of a single vector
}
\examples{
\dontrun{
res <- create_intervals(rec_2021)
summary_table(res)
}
}
\author{
Andrew Edwards
}
