% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-table.intervals_density_list.R
\name{summary_table.intervals_density_list}
\alias{summary_table.intervals_density_list}
\title{Print a Markdown summary table of main results from \code{create_intervals()}
output of a tibble (e.g. for multiple years)}
\usage{
\method{summary_table}{intervals_density_list}(int_dens_list, dig = 2)
}
\arguments{
\item{int_dens_list}{\code{intervals_density_list} object as output from
\code{create_intervals()} being applied to a tibble}

\item{dig}{number of decimal places to show}
}
\value{
Markdown code (for automatic use straight in an R Markdown document, for instance)
}
\description{
Print a Markdown summary table of main results from \code{create_intervals()}
output of a tibble (e.g. for multiple years)
}
\examples{
\dontrun{
res <- create_intervals(rec_2021) TODO
summary_table(res)
}
}
\author{
Andrew Edwards
}
