% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-table.R
\name{summary_table}
\alias{summary_table}
\title{Make \code{summary_table()} use the right method depending on class of input}
\usage{
summary_table(dat, ...)
}
\description{
Based on \code{plot()}.
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
