% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-warnings.R
\name{with_warnings}
\alias{with_warnings}
\title{With Warnings}
\usage{
with_warnings(expr)
}
\arguments{
\item{expr}{expression to be evaluated}
}
\description{
This function is used by chi.2.eksp to ensure that the function works correctly. It makes use
of functions that provide a mechanism for handling unusual conditions, including errors and warnings.
A list is returned of the value and the associated warning
}
\details{
Taken from https://rdrr.io/github/Boshoffsmit/novaReport/src/R/withWarnings.R
and then adapted. TODO update help
}
\references{
Bill Dunlap - original co-author
}
