% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-intervals.data.frame.R
\name{create_intervals.data.frame}
\alias{create_intervals.data.frame}
\title{Calculate ETIs and HDIs for a data frame of quantities.}
\usage{
\method{create_intervals}{data.frame}(dat_mcmc, allow_hdi_zero = FALSE, credibility = 0.95, ...)
}
\arguments{
\item{...}{arguments to pass to \code{create_intervals()} and then \code{density()}
TODO test changing from.}
}
\value{
list first column is quantity, then rest is like
\code{create_intervals.numeric()} TODO object of \code{res_all_years} which contains \code{year} plus
\code{intervals} and \code{dens} for that year, plus \code{intervals_all_years} which
contains all the intervals in one tibble with \code{year} as the first column.
}
\description{
For a data frame of MCMC (or other) samples as rows and some quantity
given in each named column, calculate the ETI, HDI, and kernel density for
each column, with \code{create_intervals()}.
}
\author{
Andrew Edwards
}
