% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-intervals.data.frame.R
\name{create_intervals.data.frame}
\alias{create_intervals.data.frame}
\title{Create credible intervals from a tibble or data frame}
\usage{
create_intervals.data.frame(dat)
}
\arguments{
\item{dat}{a tibble where each row represents an MCMC sample, and each
column is a quantity which we want to summarise with credible intervals.}
}
\value{
tibble with rows representing the columns of \code{dat}, and columns
\code{name} (giving each column of \code{dat}), \code{median}, \code{2.5} and \code{97.5}
representing the equal-tailed interval, \code{hdi_lower} and \code{hdi_upper}
representing the lower and upper ends of the HDI 95\% interval,
\code{width_equal} and \code{width_hdi} giving the widths of the two types of
intervals, and \code{width_diff} giving the difference between them
(\code{width_equal - width_hdi}, which must be non-negative
because HDI by definition is the narrowest interval).
}
\description{
\if{html}{\out{
 <desc>
}}
}
\examples{
\dontrun{
rec_intervals <- create_intervals(hake_recruitment_mcmc)
}
}
\author{
Andrew Edwards
}
