% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intervals.R
\name{create_intervals}
\alias{create_intervals}
\title{For a vector of values, calculate the ETI and HDI}
\usage{
create_intervals(
  dat,
  density = TRUE,
  credibility = 0.95,
  from = 0,
  n = 16384,
  tolerance = 0.01,
  ...
)
}
\arguments{
\item{dat}{numeric vector of values, such as MCMC samples for a quantity}

\item{density}{if TRUE (the default) then use the density approach for the HDI
calculation, rather than the \code{hdi()} default of just the sample values. If
FALSE then the density kernel is only used to estimate the y values of
the pdf at specified points.}

\item{credibility}{numeric value between 0 and 1 specifying the interval to
be specified (0.95 for 95\%, 0.90 for 90\%, etc.)}

\item{from}{the left-most point of the grid at which the density is to be estimated}

\item{n}{the number of equally spaced points at which the density is
to be estimated. We found the \code{density()} default of 512 to give
inaccurate results, so increased it (as a power of 2, see
\code{?density}). Changing \code{n} changes the resolution of the density kernel but
not the wiggliness.}

\item{tolerance}{amount (as a proportion) that the relative difference
between \code{y_hdi_lower} and \code{y_hdi_upper} can be, calculated as their
absolute difference divided by their mean. If the calculation is larger
than \code{tolerance} then a warning is given. Only applicable when \code{density = TRUE}.}

\item{...}{arguments to pass onto \code{density()}, including \code{to} which is
the right-most equivalent to \code{from}}
}
\value{
list object of class \code{intervals_density} (such that we can plot it
with \code{plot.intervals_density()}, with objects:
\itemize{
\item intervals: one-row tibble with columns:
\itemize{
\item median: median of the data
\item eti_lower: lower end of the ETI
\item eti_upper: upper end of the ETI
\item hdi_lower: lower end of the HDI
\item hdi_upper: upper end of the HDI
\item width_eti: width of the ETI
\item width_hdi: width of the HDI
\item width_diff: difference in widths, how much smaller (more certain) the
HDI is than the ETI
\item i_eti_lower: index for which \code{eti_lower} is between
\code{dens$x[i_eti_lower]} and \code{dens$x[i_eti_lower + 1]}
\item y_eti_lower: linearly interpolated value based on \code{i_eti_lower}
corresponding to the density at \code{eti_lower}
\item i_eti_upper, y_eti_upper: similar to \code{...lower} but for \code{upper}
\item i_hdi_lower: index for which \code{dens$x[i_hdi_lower] = hdi_lower}. The
theoretical true value of the lower bound of HDI will lie between
\code{dens$x[i_hdi_lower - 1]} and \code{dens$x[i_hdi_lower]}, but the high \code{n} used
should make this range small enough
\item y_hdi_lower: the density at \code{dens$y[i_hdi_lower]} corresponding to \code{hdi_lower}
\item i_hdi_upper: index for which \code{dens$x[i_hdi_upper] = hdi_upper}. The
theoretical true value of the upper bound of HDI will lie between
\code{dens$x[i_hdi_upper]} and \code{dens$x[i_hdi_upper + 1]} (note the asymmetry to
\code{i_hdi_lower}), but the high \code{n} used should make this range small enough
\item y_hdi_upper: the density at \code{dens$y[i_hdi_upper]} corresponding to \code{hdi_upper}
\item hdi_height: the height of the pdf returned from \code{HDInterval::hdi()},
corresponding to either \code{y_hdi_lower} or \code{y_hdi_upper} (depending on which
is the first \code{dens$x} value to push the integrated sum of the sorted
cumulative \code{dens$y} values over \code{credibility}; see
\code{HDInterval::hdi.density()}. Is \code{NA} if \code{density = FALSE}.
}
}
}
\description{
For a vector of values, calculate the ETI and HDI
}
\examples{
\dontrun{
create_intervals(rec_2021)   # rec_2021 is the MCMC samples for hake recruitment in 2021
}
}
\author{
Andrew Edwards
}
