% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure-1.R
\name{figure_1}
\alias{figure_1}
\title{Plot Figure 1 for a specific size.
\if{html}{\out{<desc>}}}
\usage{
figure_1(
  file = "figure-1.pdf",
  height = 4,
  width = 114/25.4,
  xlim = c(0, 40),
  xlab = "Recruitment (billions of fish)",
  ...
)
}
\description{
Plot Figure 1 for a specific size.
\if{html}{\out{<desc>}}
}
\examples{
\dontrun{

}
}
\author{
Andrew Edwards
}
