% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.intervals_density.R
\name{plot.intervals_density}
\alias{plot.intervals_density}
\title{Plot kernel density function of an \code{intervals_density} object, with
tails shaded as specified and ETI or HDI shown and optional explanatory lines added}
\usage{
\method{plot}{intervals_density}(
  ints_dens,
  type = "hdi",
  col_main = "blue",
  col_main_text = NULL,
  col_tail = "red",
  rug_top = FALSE,
  rug_bottom = FALSE,
  interval_arrows = FALSE,
  y_arrow = 0.098,
  arrowhead_length = 0.2,
  arrowhead_gap = 0,
  col_bars = "black",
  bars_multiplier = 1.5,
  lwd_border = 0.4,
  x_minor_ticks_by = 1,
  y_minor_ticks_by = 0.01,
  show_a_b = TRUE,
  hdi_horizontal = TRUE,
  explanatory_lines_a_b = FALSE,
  col_explanatory_lines = "darkgrey",
  explanatory_lines_extra = FALSE,
  show_discontinuity = FALSE,
  discontinuity_multiplier = 2,
  ...
)
}
\arguments{
\item{ints_dens}{object of class \code{intervals_density}, resulting from running
\code{create_intervals()} on a vector of values.}

\item{type}{type of intervals: either \code{hdi} or \code{equal}}

\item{col_main}{colour to fill in density and arrows corresponding to the interval}

\item{col_main_text}{colour to label the intervals (default is \code{col_main})}

\item{col_tail}{colour to fill in density and arrows corresponding to the tails}

\item{rug_top}{logical whether to show rug at the top for the density values}

\item{rug_bottom}{logical whether to show rug at the bottom for the raw data values}

\item{interval_arrows}{logical whether to show arrows depicting intervals}

\item{y_arrow}{value on y-axis to show the arrows depicting intervals}

\item{arrowhead_length}{length of arrowheads on intervals}

\item{arrowhead_gap}{half the gap between arrowheads of interval; manually
tweak to give a slight gap between them if they touch too
much. Half of required gap since amount is applied to each arrow.}

\item{col_bars}{colour of the bars showing regions a and b}

\item{bars_multiplier}{numeric multiplier, to nudge the bars higher (value of
1.0 puts them exactly at the minimum density of the ends of the credible
interval)}

\item{lwd_border}{line width of polygon outlines used to define the
different regions}

\item{x_minor_ticks_by, }{y_minor_ticks_by increments for adding minor tick
marks on the x and y axes}

\item{show_a_b}{logical, whether to show intervals a and b on ETI plot}

\item{hdi_horizontal}{logical, whether to horizontal line at density of HDI}

\item{explanatory_lines_a_b}{logical, whether to plot extra lines to explain the
ranges a and b for the ETI.}

\item{col_explanatory_lines}{colour for explanatory lines}

\item{explanatory_lines_extra}{logical, whether to plot extra lines to explain the
ranges c and d for which all values in c are more likely than those in d,
yet c is outside the ETI and d is inside it; see vignette.}

\item{show_discontinuity}{logical, if \code{TRUE} then plot the discontinuity in
the HDI that arises (only matters if \code{ints_dens$intervals$warning == TRUE}) as points.}

\item{discontinuity_multiplier}{how much to move the \code{show_discontinuity}
points up from the density level of the HDI (just to improve how it looks)}

\item{...}{arguments to pass onto \code{plot()}}
}
\value{
The desired plot.
}
\description{
Creates plot on object of class \code{intervals_density}, which is the
result of running \code{create_intervals()} on a vector of values. Used to make
Fig. 1A. See vignettes. For left-skewed distributions, the extra
functionality that adds explanatory lines etc. is not fully implemented, but
the basic figure works okay. Please make an Issue on GitHub if you would
find it helpful for this to be fixed (it just seemed a low priority).
}
\examples{
\dontrun{
res <- create_intervals(rec_2021)
plot(res)

# Left-skewed example:
left <- 60 - rec_2021[rec_2021 < 60]
res <- create_intervals(left)
par(mfrow = c(2, 1))
plot(res, type = "eti", show_a_b = FALSE)
plot(res, type = "hdi", show_a_b = FALSE)
summary_table(res)    # shows that a and b ranges are not properly
  calculated yet for left-skewed distributions. Create an Issue if you would
  like this functionality.
# And see results and result-extra vignettes, rendered on the GitHub page.
}
}
\author{
Andrew Edwards
}
