% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.intervals_density_list.R
\name{plot.intervals_density_list}
\alias{plot.intervals_density_list}
\title{Plot a time series of intervals from multiple samples, as calculated from
\code{create_intervals()} applied to a data frame object.}
\usage{
\method{plot}{intervals_density_list}(
  obj,
  type = "comparison",
  eti_bar_col = "blue",
  hdi_bar_col = "red",
  add_line_at_0.4 = FALSE,
  add_line_at_0.4_col = "darkgreen",
  add_line_at_0.4_lty = 5,
  inc = 0.15,
  add_big_ticks_x = TRUE,
  ylim = NULL,
  y_tick_start = 0,
  y_tick_end = NULL,
  y_tick_by = 1,
  pch = 20,
  cex = 0.8,
  add_legend = TRUE,
  leg_loc = "topright",
  inset = c(0, -0.02),
  join_intervals = FALSE,
  arrowhead_length = 0.15,
  mfrow = c(1, 1),
  ...
)
}
\arguments{
\item{obj}{\code{intervals_density_list} object from running \code{\link[=create_intervals]{create_intervals()}}
on a data frame, see full description in the return section of \code{?create_intervals}}

\item{type}{\code{comparison} to plot both intervals for each value of \code{quantity}
(such as year, thus giving a time series, only works if \code{quantity} is
numeric), or \code{eti} or \code{hdi} to plot the
respective full distribution  and interval for each value of \code{quantity};
\code{quantity} is the first column of \code{obj$intervals_all_years}, corresponding
to the original column names of the data frame used to crate \code{obj}.}

\item{eti_bar_col}{colour used to plot the ETIs in comparison plot}

\item{hdi_bar_col}{colour used to plot the HDIs in comparison plot}

\item{add_line_at_0.4}{logical whether to add a line at 0.4, specifically
for Figure 2B, also adds the 'Now' and 'Projections' text}

\item{add_line_at_0.4_col}{colour for 0.4 line}

\item{add_line_at_0.4_lty}{linetype for 0.4 line}

\item{inc}{amount to shift ETI to left and HDI to right in comparison, tweak
as appropriate (depends on scale of x-axis)}

\item{add_big_ticks_x}{slightly larger ticks every 10 values of the x-axis
quantity, since not all get labelled automatically}

\item{ylim}{range for y-axis, if \code{NULL} (the default) then created
automatically (can probably subsume into \code{...});}

\item{y_tick_start}{where to start y tickmarks}

\item{y_tick_end}{where to end y tickmarks}

\item{y_tick_by}{increment for y tickmarks}

\item{pch}{style of points for medians for comparison plots}

\item{cex}{size of points for medians for comparison plots; set to 0 to not
show medians}

\item{add_legend}{logical, whether to add a legend to comparison plot}

\item{leg_loc}{location of legend, passed onto \code{\link[=legend]{legend()}} as the \code{legend}
argument}

\item{inset, }{vector of length two for shifting the legend (\code{inset} argument
to \code{\link[=legend]{legend()}}.}

\item{join_intervals}{logical, if \code{TRUE} then join up the ends of the
intervals, which can be visually useful}

\item{arrowhead_length}{length of arrowhead, specifically for the
\code{add_line_at_0.4} relative biomass plot}

\item{mfrow}{\code{par(mfrow)} vector of length two (number of rows and number of
columns to plot panels into) to pass into \code{\link[=plot_densities]{plot_densities()}}}

\item{...}{further arguments passed onto [\code{\link[=plot.default]{plot.default()}} for \code{type == "comparison"}, else passed onto \code{\link[=plot_densities]{plot_densities()}} for \code{type} being \code{eti}
or \code{hdi} (and then onto \code{\link[=plot.intervals_density]{plot.intervals_density()}}).}
}
\value{
plot of the required style (multiple plots for \code{type} being \code{eti} or
\code{hdi}
}
\description{
Default plot is to show show ETIs, HDIs, or both, as saved in the
\code{intervals_all} component of the list from
\code{create_intervals()}. See example and vignette. Also includes option to plot
the density for each quantity (years in ##' our recruitment example).
}
\examples{
\dontrun{
# And see vignettes
res_all_years <- create_intervals(dplyr::select(hake_recruitment_mcmc,
                                                         -"Virgin"))     # TODO
plot(res_all_years)
}
}
\author{
Andrew Edwards
}
