% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-densitites.R
\name{plot_densities}
\alias{plot_densities}
\title{Plots densities and ETIs or HDIs for multiplie quantities (e.g. years).}
\usage{
plot_densities(obj, type, quantity = NULL, mfrow = c(1, 1), ...)
}
\description{
Plots the density and ETI/HDI for each quantitiey (such as year) in a
\code{intervals_density_list} object that is returned from \code{\link[=create_intervals]{create_intervals()}}
applied to a data frame. Called from \code{\link[=plot.intervals_density_list]{plot.intervals_density_list()}},
probably not needed by users.
}
\author{
Andrew Edwards
}
